+++ crates/ruma-state-res/src/lib.rs
@@ -466,12 +466,19 @@ fn iterative_auth_check<E: Event + Clone>(
             }
         }
 
-        if auth_check(room_version, &event, |ty, key| auth_events.get(&ty.with_state_key(key)))? {
-            // add event to resolved state map
-            resolved_state.insert(event.event_type().with_state_key(state_key), event_id.clone());
-        } else {
-            // synapse passes here on AuthError. We do not add this event to resolved_state.
-            warn!("event failed the authentication check");
+        match auth_check(room_version, &event, |ty, key| auth_events.get(&ty.with_state_key(key))) {
+            Ok(true) => {
+                // Add event to resolved state.
+                resolved_state
+                    .insert(event.event_type().with_state_key(state_key), event_id.clone());
+            }
+            // Don't add this event to resolved_state.
+            Ok(false) => {
+                warn!("event failed the authentication check");
+            }
+            Err(error) => {
+                warn!("event failed the authentication check: {error}");
+            }
         }
 
         // TODO: if these functions are ever made async here